- Modifying lib name to metadata and organizing inner folders - Solving broken imports after refactor
