+++ demo/components/LiveSchemaEditor.js
@@ -10,8 +10,8 @@ import {Alert, Badge} from 'react-bootstrap';
 import moment from 'moment';
 import numbro from 'numbro';
 import reactWidgetsMomentLocalizer from 'react-widgets/lib/localizers/moment';
-import momentLocalizer from '../../src/lib/localization/momentLocalizer';
-import numbroLocalizer from '../../src/lib/localization/numbroLocalizer';
+import momentLocalizer from '../../src/localization/momentLocalizer';
+import numbroLocalizer from '../../src/localization/numbroLocalizer';
 import ButtonToolbar from './ButtonToolbar';
 import FormOptions from './FormOptions';
