+++ src/localization/numberLocalizer.js
