+++ src/metadata/evaluator/property-filter/defaultMetadataPropertyFilter.js
