+++ src/metadata/helper/arrayHelper.js
