+++ src/metadata/helper/expressionHelper.js
@@ -1,4 +1,4 @@
-import { getDateLocalizer } from '../localization/dateLocalizer';
+import { getDateLocalizer } from '../../localization/dateLocalizer';
 import formato from 'formato';
 
 /**
