+++ src/metadata/model/parsers/arrayParser.js
