+++ src/metadata/model/parsers/dateTimeParser.js
@@ -1,5 +1,6 @@
-import { getFormat } from '../helpers/dateTimeHelpers';
-import { getDateLocalizer } from '../localization/dateLocalizer';
+import { getFormat } from '../../helper/dateTimeHelpers';
+import { getDateLocalizer } from '../../../localization/dateLocalizer';
+
 var localizer;
 
 export default function parse(metadata, value) {
