+++ src/metadata/model/parsers/entityParser.js
