+++ src/metadata/model/processor/numberProcessor.js
@@ -1,4 +1,4 @@
-import { getNumberLocalizer } from '../localization/numberLocalizer';
+import { getNumberLocalizer } from '../../../localization/numberLocalizer';
 var localizer;
 
 export default function parse(metadata, value) {
