+++ src/metadata/validator/validators/arrayValidator.js
