+++ src/metadata/validator/validators/requiredValidator.js
@@ -1,4 +1,4 @@
-import expressionEvaluator from '../expressionEvaluator';
+import expressionEvaluator from '../../evaluator/expressionEvaluator';
 
 /**
  * Validates required properties
