+++ test/index.js
@@ -1,8 +1,8 @@
 import 'es5-shim';
 import moment from 'moment';
-import momentLocalizer from '../src/lib/localization/momentLocalizer';
+import momentLocalizer from '../src/localization/momentLocalizer';
 import numbro from 'numbro';
-import numbroLocalizer from '../src/lib/localization/numbroLocalizer';
+import numbroLocalizer from '../src/localization/numbroLocalizer';
 
 beforeEach(function() {
     sinon.stub(console, 'warn');
